requirements.txt
Pillow==10.4.0
segno==1.6.1